# Correctly emulate FPU concurrent execution timings