--- a/src/cpu/386.c
+++ b/src/cpu/386.c
@@ -81,7 +81,20 @@ x386_log(const char *fmt, ...)
 
 #define OP_TABLE(name) ops_ ## name
 
-#define CLOCK_CYCLES(c) cycles -= (c)
+#define CLOCK_CYCLES(c) \
+	{\
+		if (fpu_cycles > 0) {\
+			fpu_cycles -= (c);\
+			if (fpu_cycles < 0) {\
+				cycles += fpu_cycles;\
+			}\
+		} else {\
+			cycles -= (c);\
+		}\
+	}
+
+#define CLOCK_CYCLES_FPU(c) cycles -= (c)
+#define CONCURRENCY_CYCLES(c) fpu_cycles = (c)
 #define CLOCK_CYCLES_ALWAYS(c) cycles -= (c)
 
 #include "x86_ops.h"