--- a/src/cpu/386_dynarec.c
+++ b/src/cpu/386_dynarec.c
@@ -267,8 +267,21 @@ static void prefetch_flush()
 
 
 #define OP_TABLE(name) ops_ ## name
-#define CLOCK_CYCLES(c) cycles -= (c)
+#define CLOCK_CYCLES(c) \
+	{\
+		if (fpu_cycles > 0) {\
+			fpu_cycles -= (c);\
+			if (fpu_cycles < 0) {\
+				cycles += fpu_cycles;\
+			}\
+		} else {\
+			cycles -= (c);\
+		}\
+	}
+#define CLOCK_CYCLES_FPU(c) cycles -= (c)
+#define CONCURRENCY_CYCLES(c) fpu_cycles = (c)
 #define CLOCK_CYCLES_ALWAYS(c) cycles -= (c)
+
 
 #include "386_ops.h"