--- a/src/cpu/x87_ops_loadstore.h
+++ b/src/cpu/x87_ops_loadstore.h
@@ -23,7 +23,8 @@ static int opFILDiw_a16(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         temp = geteaw(); if (cpu_state.abrt) return 1;
         x87_push((double)temp);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fild_16) : (x87_timings.fild_16 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fild_16) : (x87_timings.fild_16 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fild_16) : (x87_concurrency.fild_16 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -35,7 +36,8 @@ static int opFILDiw_a32(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         temp = geteaw(); if (cpu_state.abrt) return 1;
         x87_push((double)temp);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fild_16) : (x87_timings.fild_16 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fild_16) : (x87_timings.fild_16 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fild_16) : (x87_concurrency.fild_16 * cpu_multi));
         return 0;
 }
 #endif
@@ -46,7 +48,8 @@ static int opFISTiw_a16(uint32_t fetchdat)
         fetch_ea_16(fetchdat);
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteaw(x87_fround16(ST(0)));
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_16) : (x87_concurrency.fist_16 * cpu_multi));
         return cpu_state.abrt;
 }
 #ifndef FPU_8087
@@ -56,7 +59,8 @@ static int opFISTiw_a32(uint32_t fetchdat)
         fetch_ea_32(fetchdat);
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteaw(x87_fround16(ST(0)));
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_16) : (x87_concurrency.fist_16 * cpu_multi));
         return cpu_state.abrt;
 }
 #endif
@@ -68,7 +72,8 @@ static int opFISTPiw_a16(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteaw(x87_fround16(ST(0))); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_16) : (x87_concurrency.fist_16 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -79,7 +84,8 @@ static int opFISTPiw_a32(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteaw(x87_fround16(ST(0))); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_16) : (x87_timings.fist_16 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_16) : (x87_concurrency.fist_16 * cpu_multi));
         return 0;
 }
 #endif
@@ -95,7 +101,8 @@ static int opFILDiq_a16(uint32_t fetchdat)
         cpu_state.MM[cpu_state.TOP&7].q = temp64;
 	FP_TAG_DEFAULT;
 
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fild_64) : (x87_timings.fild_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fild_64) : (x87_timings.fild_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fild_64) : (x87_concurrency.fild_64 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -110,7 +117,8 @@ static int opFILDiq_a32(uint32_t fetchdat)
         cpu_state.MM[cpu_state.TOP&7].q = temp64;
 	FP_TAG_DEFAULT;
 
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fild_64) : (x87_timings.fild_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fild_64) : (x87_timings.fild_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fild_64) : (x87_concurrency.fild_64 * cpu_multi));
         return 0;
 }
 #endif
@@ -139,7 +147,8 @@ static int FBSTP_a16(uint32_t fetchdat)
         if (ST(0) < 0.0) tempc |= 0x80;
         writememb(easeg, cpu_state.eaaddr + 9, tempc); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fbstp) : (x87_timings.fbstp * cpu_multi));		
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fbstp) : (x87_timings.fbstp * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fbstp) : (x87_concurrency.fbstp * cpu_multi));		
         return 0;
 }
 #ifndef FPU_8087
@@ -167,7 +176,7 @@ static int FBSTP_a32(uint32_t fetchdat)
         if (ST(0) < 0.0) tempc |= 0x80;
         writememb(easeg, cpu_state.eaaddr + 9, tempc); if (cpu_state.abrt) return 1;
         x87_pop();
-		CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fbstp) : (x87_timings.fbstp * cpu_multi));
+		CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fbstp) : (x87_timings.fbstp * cpu_multi));
         return 0;
 }
 #endif
@@ -184,7 +193,8 @@ static int FISTPiq_a16(uint32_t fetchdat)
                 temp64 = x87_fround(ST(0));
         seteaq(temp64); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_64) : (x87_timings.fist_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_64) : (x87_timings.fist_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_64) : (x87_concurrency.fist_64 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -200,7 +210,8 @@ static int FISTPiq_a32(uint32_t fetchdat)
                 temp64 = x87_fround(ST(0));
         seteaq(temp64); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_64) : (x87_timings.fist_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_64) : (x87_timings.fist_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_64) : (x87_concurrency.fist_64 * cpu_multi));
         return 0;
 }
 #endif
@@ -213,7 +224,8 @@ static int opFILDil_a16(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         templ = geteal(); if (cpu_state.abrt) return 1;
         x87_push((double)templ);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fild_32) : (x87_timings.fild_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fild_32) : (x87_timings.fild_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fild_32) : (x87_concurrency.fild_32 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -225,7 +237,8 @@ static int opFILDil_a32(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         templ = geteal(); if (cpu_state.abrt) return 1;
         x87_push((double)templ);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fild_32) : (x87_timings.fild_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fild_32) : (x87_timings.fild_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fild_32) : (x87_concurrency.fild_32 * cpu_multi));
         return 0;
 }
 #endif
@@ -236,7 +249,8 @@ static int opFISTil_a16(uint32_t fetchdat)
         fetch_ea_16(fetchdat);
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteal(x87_fround32(ST(0)));
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_32) : (x87_concurrency.fist_32 * cpu_multi));
         return cpu_state.abrt;
 }
 #ifndef FPU_8087
@@ -246,7 +260,8 @@ static int opFISTil_a32(uint32_t fetchdat)
         fetch_ea_32(fetchdat);
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteal(x87_fround32(ST(0)));
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_32) : (x87_concurrency.fist_32 * cpu_multi));
         return cpu_state.abrt;
 }
 #endif
@@ -258,7 +273,8 @@ static int opFISTPil_a16(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteal(x87_fround32(ST(0))); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_32) : (x87_concurrency.fist_32 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -269,7 +285,8 @@ static int opFISTPil_a32(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         seteal(x87_fround32(ST(0))); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fist_32) : (x87_timings.fist_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fist_32) : (x87_concurrency.fist_32 * cpu_multi));
         return 0;
 }
 #endif
@@ -282,7 +299,8 @@ static int opFLDe_a16(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         t=x87_ld80(); if (cpu_state.abrt) return 1;
         x87_push(t);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fld_80) : (x87_concurrency.fld_80 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -294,7 +312,8 @@ static int opFLDe_a32(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         t=x87_ld80(); if (cpu_state.abrt) return 1;
         x87_push(t);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fld_80) : (x87_concurrency.fld_80 * cpu_multi));
         return 0;
 }
 #endif
@@ -306,7 +325,8 @@ static int opFSTPe_a16(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         x87_st80(ST(0)); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fld_80) : (x87_concurrency.fld_80 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -317,7 +337,8 @@ static int opFSTPe_a32(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         x87_st80(ST(0)); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fld_80) : (x87_timings.fld_80 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fld_80) : (x87_concurrency.fld_80 * cpu_multi));
         return 0;
 }
 #endif
@@ -330,7 +351,8 @@ static int opFLDd_a16(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         t.i = geteaq(); if (cpu_state.abrt) return 1;
         x87_push(t.d);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fld_64) : (x87_timings.fld_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fld_64) : (x87_timings.fld_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fld_64) : (x87_concurrency.fld_64 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -342,7 +364,8 @@ static int opFLDd_a32(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         t.i = geteaq(); if (cpu_state.abrt) return 1;
         x87_push(t.d);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fld_64) : (x87_timings.fld_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fld_64) : (x87_timings.fld_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fld_64) : (x87_concurrency.fld_64 * cpu_multi));
         return 0;
 }
 #endif
@@ -355,7 +378,8 @@ static int opFSTd_a16(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         t.d = ST(0);
         seteaq(t.i);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_64) : (x87_concurrency.fst_64 * cpu_multi));
         return cpu_state.abrt;
 }
 #ifndef FPU_8087
@@ -367,7 +391,8 @@ static int opFSTd_a32(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         t.d = ST(0);
         seteaq(t.i);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_64) : (x87_concurrency.fst_64 * cpu_multi));
         return cpu_state.abrt;
 }
 #endif
@@ -381,7 +406,8 @@ static int opFSTPd_a16(uint32_t fetchdat)
         t.d = ST(0);
         seteaq(t.i); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_64) : (x87_concurrency.fst_64 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -394,7 +420,8 @@ static int opFSTPd_a32(uint32_t fetchdat)
         t.d = ST(0);
         seteaq(t.i); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_64) : (x87_timings.fst_64 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_64) : (x87_concurrency.fst_64 * cpu_multi));
         return 0;
 }
 #endif
@@ -407,7 +434,8 @@ static int opFLDs_a16(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         ts.i = geteal(); if (cpu_state.abrt) return 1;
         x87_push((double)ts.s);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_32) : (x87_concurrency.fst_32 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -419,7 +447,8 @@ static int opFLDs_a32(uint32_t fetchdat)
 	SEG_CHECK_READ(cpu_state.ea_seg);
         ts.i = geteal(); if (cpu_state.abrt) return 1;
         x87_push((double)ts.s);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_32) : (x87_concurrency.fst_32 * cpu_multi));
         return 0;
 }
 #endif
@@ -432,7 +461,8 @@ static int opFSTs_a16(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         ts.s = (float)ST(0);
         seteal(ts.i);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_32) : (x87_concurrency.fst_32 * cpu_multi));
         return cpu_state.abrt;
 }
 #ifndef FPU_8087
@@ -444,7 +474,8 @@ static int opFSTs_a32(uint32_t fetchdat)
 	SEG_CHECK_WRITE(cpu_state.ea_seg);
         ts.s = (float)ST(0);
         seteal(ts.i);
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_32) : (x87_concurrency.fst_32 * cpu_multi));
         return cpu_state.abrt;
 }
 #endif
@@ -458,7 +489,8 @@ static int opFSTPs_a16(uint32_t fetchdat)
         ts.s = (float)ST(0);
         seteal(ts.i); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_32) : (x87_concurrency.fst_32 * cpu_multi));
         return 0;
 }
 #ifndef FPU_8087
@@ -471,7 +503,8 @@ static int opFSTPs_a32(uint32_t fetchdat)
         ts.s = (float)ST(0);
         seteal(ts.i); if (cpu_state.abrt) return 1;
         x87_pop();
-        CLOCK_CYCLES((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CLOCK_CYCLES_FPU((fpu_type >= FPU_487SX) ? (x87_timings.fst_32) : (x87_timings.fst_32 * cpu_multi));
+        CONCURRENCY_CYCLES((fpu_type >= FPU_487SX) ? (x87_concurrency.fst_32) : (x87_concurrency.fst_32 * cpu_multi));
         return 0;
 }
 #endif