… src/cpu/x87_ops_arith.h
… src/cpu/x87_ops_misc.h
… src/cpu/x87_timings.c